refactor: view 결합도 감소
