+++ src/main/java/roomescape/reservation/dto/ReservationResponse.java
@@ -1,0 +1,15 @@
+package roomescape.reservation.dto;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import roomescape.reservation.Reservation;
+
+public record ReservationResponse(long id, String name, LocalDate date, LocalTime time) {
+    public static ReservationResponse from(Reservation reservation) {
+        return new ReservationResponse(
+                reservation.getId(),
+                reservation.getName(),
+                reservation.getDate(),
+                reservation.getTime());
+    }
+}
